requirements.txt
fastapi
therapist
mypy
bandit
black
isort
autoflake
uvicorn[standard]
auth0-python
pytest
toml
twilio